transitions between poses
add a set of target angles to the model to record how the model is
supposed to be posed now.
refactor old pose() method to set target angles for each pose instead
of actually setting angles in the model
add updatePose() method to smoothly move the model from the current
position to the target position.
rename old update() to updateVelocity()